More miscellaneous work on getting physics to a functional stage.

+++ src/guiElement.h
@@ -51,7 +51,7 @@ void guiElementInit(guiElement *const restrict gui, const guiElementType_t type)
 
 void guiElementUpdate(guiElement *const restrict gui, const float time);
 void guiElementDraw(
-	guiElement *const restrict gui, const float windowWidth, const float windowHeight, const shaderSprite *const restrict shader
+	guiElement *const restrict gui, const int windowWidth, const int windowHeight, const shaderSprite *const restrict shader
 );
 
 void guiElementDelete(guiElement *const restrict gui);
